refactor: simplifying expression and event handler
