+++ src/test/kotlin/at/ac/uibk/dps/cirrina/EventsTest.kt
@@ -1,12 +1,12 @@
 package at.ac.uibk.dps.cirrina
 
-import EventHandlerInMemory
 import at.ac.uibk.dps.cirrina.csm.Csml
 import at.ac.uibk.dps.cirrina.csm.Csml.EventChannel
 import at.ac.uibk.dps.cirrina.data.DefaultDescriptions
 import at.ac.uibk.dps.cirrina.di.DaggerTestComponent
 import at.ac.uibk.dps.cirrina.di.TestModule
 import at.ac.uibk.dps.cirrina.execution.`object`.Event
+import at.ac.uibk.dps.cirrina.execution.`object`.EventHandler
 import at.ac.uibk.dps.cirrina.execution.`object`.EventHandler.Companion.PERIPHERAL_SOURCE
 import at.ac.uibk.dps.cirrina.execution.provider.ContextInMemory
 import java.time.Duration
@@ -25,7 +25,7 @@ class EventsTest {
   @Test
   fun testEventsExecute() = runBlocking {
     assertDoesNotThrow {
-      val eventHandler = EventHandlerInMemory()
+      val eventHandler = EventHandler()
       val context = ContextInMemory()
 
       val runtime =
